requirements/base.txt
Django==2.0.7
django-attachments>=1.3
django-grappelli
django-vinaigrette
django-uuslug
odfpy
python-bugzilla
# because of https://github.com/pycontribs/jira/issues/501
jira==1.0.10
PyGithub
dj-pagination
django-tinymce
beautifulsoup4
django-contrib-comments
html2text
django-modern-rpc
